+++ db/migrate/002_add_default_mentions.rb
@@ -1,0 +1,7 @@
+# Redmine Messenger plugin for Redmine
+
+class AddDefaultMentions < ActiveRecord::Migration
+  def change
+    add_column :messenger_settings, :default_mentions, :string
+  end
+end